Update documentation and package arrangement
This adds documentation to every package, module and function, and
updates some which were a little out-of-date.

As part of updating the documentation, I also rearranged some functions
in some modules, so the module structure is a little bit more
self-documenting, and not so spread out over several different files
that lal do largely the same thing.